feat: Add `is_mezmo_metric()` function
Adds a new function that checks the structure of a provided value and
determines if it's in the Mezmo metric format.

Example:

```vrl
is_metric, err = is_mezmo_metric(
  {
    "kind":"incremental",
    "name":"metric",
    "value":{"type":"counter","value":1}
  }
)

if is_metric {
  log("It's a metric!")
} else {
  log(err)
}
```

`err` contains the reason a value is not considered a metric. This can
be use to provide the reason it failed or it can be ignored.

Ref: LOG-16685