feat(lookup): exposing vector lookup logic through wasm
Expose two simple methods - one for parsing VRL paths into segments and
one for joining segments into a VRL path. This accounts for for fields,
indexes and coalesce. The code will also handle quoting and minimizing
quoting on field to be more human friendly.

These two functions could be used to perform richer validation on user
input and simplify some of the more complex path construction in the
pipeline service and/or front end UI.

Ref: LOG-17092, LOG-16869