feat(mezmo_map_key_values): Implement mezmo map key values function
Implement mezmo function for mapping an object
or array of items. In the case of objects, it will iterate
the keys and values. For arrays, it will iterate the
entries of the array. Object keys may be modified in the
provided closure. Object and array values may also be modified
in the closure